refactor(workspace): share Next API plumbing via factory (P5.5)
Final P5 slice — the request/auth/error/validation plumbing moves to
packages/shared/src/api (@babyjamjam/shared/api) behind a
createRouteUtils({ serverAPIClient, secureCookies, errorResponseMode })
factory, removing any shared->app-local coupling (env, API client).
Both apps keep thin binding shims at the old @/lib/api/route-utils
path — zero route-file churn.

Mobile's hardened behavior (parseBody/zod, bodySchema proxy helpers,
sanitized upstream errors) is the default; frontend keeps its legacy
errorResponse contract via errorResponseMode and GAINS the zod plumbing
for future retrofits. Routes stay per-app by design. zod becomes a
shared-package dependency at the locked version.

Gates: both type-checks clean, 632 + 245 tests green (mobile's route
tests exercise the plumbing unchanged), lint counts unchanged, both
production builds exit 0.
