+++ packages/shared/package.json
@@ -8,6 +8,7 @@
     "types": "src/index.ts",
     "exports": {
         ".": "./src/index.ts",
+        "./api": "./src/api/index.ts",
         "./types/alimtalk": "./src/types/alimtalk.ts",
         "./types/eformsign": "./src/types/eformsign.ts",
         "./types/message": "./src/types/message.ts",
@@ -20,7 +21,8 @@
     },
     "dependencies": {
         "clsx": "^2.1.1",
-        "tailwind-merge": "^3.4.0"
+        "tailwind-merge": "^3.4.0",
+        "zod": "^4.3.6"
     },
     "devDependencies": {
         "typescript": "^5"
